+++ contrib/parseq-restli-client/src/main/java/com/linkedin/restli/client/config/RequestConfigProviderBuilder.java
@@ -4,21 +4,21 @@ import com.linkedin.restli.client.InboundRequestContextFinder;
 import com.linkedin.restli.client.ParSeqRestClientConfig;
 import com.linkedin.restli.client.ParSeqRestClientConfigBuilder;
 
-class ResourceConfigProviderBuilder {
+class RequestConfigProviderBuilder {
 
   private final ParSeqRestClientConfigBuilder _config = new ParSeqRestClientConfigBuilder();
   private InboundRequestContextFinder _inboundRequestFinder;
 
-  public ResourceConfigProvider build() throws ResourceConfigKeyParsingException {
-    return new ResourceConfigProviderImpl(_inboundRequestFinder, _config.build());
+  public RequestConfigProvider build() throws RequestConfigKeyParsingException {
+    return new RequestConfigProviderImpl(_inboundRequestFinder, _config.build());
   }
 
-  public ResourceConfigProviderBuilder addConfig(ParSeqRestClientConfig config) {
+  public RequestConfigProviderBuilder addConfig(ParSeqRestClientConfig config) {
     _config.addConfig(config);
     return this;
   }
 
-  public ResourceConfigProviderBuilder setInboundRequestFinder(InboundRequestContextFinder inboundRequestFinder) {
+  public RequestConfigProviderBuilder setInboundRequestFinder(InboundRequestContextFinder inboundRequestFinder) {
     _inboundRequestFinder = inboundRequestFinder;
     return this;
   }